Moved gulp verifier functionality into another file inside gulp since it will get too large
